requirements.txt
Django==4.2.8
djangorestframework==3.14.0
requests==2.31.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.0.3
asgiref==3.7.0
backports.zoneinfo==0.2.1;python_version<"3.9"
Brotli==1.0.9
gunicorn==20.1.0
sqlparse==0.4.4
typing-extensions==4.6.0
whitenoise==6.4.0
